agents/pyproject.toml
[project]
name = "management-agent"
version = "0.1.0"
description = "Management Agent template for LangSmith/LangGraph deployment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.4.0",
    "fastapi>=0.135.3",
    "langgraph>=1.0.0",
    "langgraph-checkpoint-postgres>=3.0.5",
    "langgraph-cli[inmem]>=0.4.19",
    "langgraph-sdk>=0.1.0",
    "langsmith>=0.3.0",
    "psycopg>=3.3.3",
    "python-dotenv>=1.0.1",
    "shared",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "pytest>=8.3.5",
    "ruff>=0.8.0",
    "shared"
]

[tool.uv.build-backend]
module-name = "management_agent"
source-include = ["langgraph.json", ".env", "Makefile"]

[tool.uv.sources]
shared = { workspace = true }
